User only retrieves his own data

+++ src/main/java/br/dev/mhc/financialassistantapi/services/UserService.java
@@ -9,6 +9,7 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,13 +17,24 @@ import org.springframework.transaction.annotation.Transactional;
 import br.dev.mhc.financialassistantapi.dto.UserDTO;
 import br.dev.mhc.financialassistantapi.dto.UserNewDTO;
 import br.dev.mhc.financialassistantapi.entities.User;
+import br.dev.mhc.financialassistantapi.entities.enums.Profile;
 import br.dev.mhc.financialassistantapi.repositories.UserRepository;
+import br.dev.mhc.financialassistantapi.security.UserSpringSecurity;
 import br.dev.mhc.financialassistantapi.services.email.EmailService;
+import br.dev.mhc.financialassistantapi.services.exceptions.AuthorizationException;
 import br.dev.mhc.financialassistantapi.services.exceptions.DataIntegrityException;
 import br.dev.mhc.financialassistantapi.services.exceptions.ObjectNotFoundException;
 
 @Service
 public class UserService {
+
+	public static UserSpringSecurity authenticate() {
+		try {
+		return (UserSpringSecurity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		} catch (Exception e) {
+			return null;
+		}
+	}
 
 	@Autowired
 	private BCryptPasswordEncoder pe;
@@ -46,6 +58,11 @@ public class UserService {
 
 	@Transactional
 	public User update(User obj) {
+		UserSpringSecurity userSS = UserService.authenticate();
+		if(userSS==null || !userSS.hasRole(Profile.ADMIN) && !obj.getId().equals(userSS.getId())) {
+			throw new AuthorizationException("Access denied");
+		}
+
 		User newObj = findById(obj.getId());
 		updateData(newObj, obj);
 		return repository.save(newObj);
@@ -57,6 +74,11 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
+		UserSpringSecurity userSS = UserService.authenticate();
+		if(userSS==null || !userSS.hasRole(Profile.ADMIN) && !id.equals(userSS.getId())) {
+			throw new AuthorizationException("Access denied");
+		}
+
 		findById(id);
 		try {
 			repository.deleteById(id);
@@ -75,6 +97,11 @@ public class UserService {
 	}
 
 	public User findById(Long id) {
+		UserSpringSecurity userSS = UserService.authenticate();
+		if(userSS==null || !userSS.hasRole(Profile.ADMIN) && !id.equals(userSS.getId())) {
+			throw new AuthorizationException("Access denied");
+		}
+		
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(
 				() -> new ObjectNotFoundException("Object not found! Id: " + id + ", Type: " + User.class.getName()));
@@ -92,7 +119,6 @@ public class UserService {
 	public User fromDTO(UserNewDTO objDTO) {
 		User user = new User(null, objDTO.getNickname(), objDTO.getEmail(), pe.encode(objDTO.getPassword()), null, null,
 				true);
-		
 		return user;
 	}
 }
